Added ContainerCreateEvent, refactor event hiearchy
