+++ Ductus.FluentDocker/Model/Events/ImagePullEvent.cs
@@ -1,0 +1,28 @@
+namespace Ductus.FluentDocker.Model.Events
+{
+  public sealed class ImagePullEvent : FdEvent<ImagePullEvent.ImagePullActor>
+  {
+    /// <summary>
+    /// Emitted when a remote image has been pulled onto local store.
+    /// </summary>
+    public ImagePullEvent()
+    {
+      Action = EventAction.Pull;
+      Type = EventType.Image;
+    }
+
+    /// <summary>
+    /// The actor is the image name and label
+    /// </summary>
+    /// <remarks>
+    /// The <see cref="EventActor.Id"/> contain the "image name:label".
+    /// </remarks>
+    public sealed class ImagePullActor : EventActor
+    {
+      /// <summary>
+      /// Name of the image without label.
+      /// </summary>
+      public string Name { get; set; }
+    }
+  }
+}
