--- Ductus.FluentDocker/Model/HostEvents/EventActor.cs
@@ -1,7 +1,0 @@
-namespace Ductus.FluentDocker.Model.HostEvents
-{
-  public class EventActor
-  {
-    public string Id { get; set; }
-  }
-}
