--- Ductus.FluentDocker/Model/HostEvents/EventScope.cs
@@ -1,8 +1,0 @@
-namespace Ductus.FluentDocker.Model.HostEvents
-{
-  public enum EventScope
-  {
-    Unknown,
-    Local
-  }
-}
